chore : change directory structure
