+++ src/main/java/com/Techeer/Team_C/User/dto/LoginFormDto.java
@@ -1,16 +1,15 @@
-package com.Techeer.Team_C.dto;
+package com.Techeer.Team_C.User.dto;
 
 import javax.validation.constraints.NotEmpty;
 
 
 
-public class SignupFormDto {
+public class LoginFormDto {
     @NotEmpty(message ="id값은 필수사항 입니다.")
     private String userId;
     @NotEmpty(message ="비밀번호 값은 필수사항 입니다.")
     private String password;
-    @NotEmpty(message ="닉네임 값은 필수사항 입니다.")
-    private String userName;
+
 
     public String getUserId() {
         return userId;
@@ -27,15 +26,6 @@ public class SignupFormDto {
     public void setPassword(String password) {
         this.password = password;
     }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
 
 
 }
