+++ src/main/java/com/Techeer/Team_C/User/dto/SignupFormDto.java
@@ -1,15 +1,16 @@
-package com.Techeer.Team_C.dto;
+package com.Techeer.Team_C.User.dto;
 
 import javax.validation.constraints.NotEmpty;
 
 
 
-public class LoginFormDto {
+public class SignupFormDto {
     @NotEmpty(message ="id값은 필수사항 입니다.")
     private String userId;
     @NotEmpty(message ="비밀번호 값은 필수사항 입니다.")
     private String password;
-
+    @NotEmpty(message ="닉네임 값은 필수사항 입니다.")
+    private String userName;
 
     public String getUserId() {
         return userId;
@@ -26,6 +27,15 @@ public class LoginFormDto {
     public void setPassword(String password) {
         this.password = password;
     }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+
 
 
 }
