+++ src/main/java/com/Techeer/Team_C/User/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.Techeer.Team_C.dto;
+package com.Techeer.Team_C.User.dto;
 //DB의 테이블구조와 직접적으로 연관된 entity를 사용하는것은 좋지 않기때문에 중간의 데이터 교환을 위한 Dto가 필요하다.
 // entity의 속성과 동일하게 작성한다.
