+++ src/main/java/com/Techeer/Team_C/User/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package com.Techeer.Team_C.repository;
+package com.Techeer.Team_C.User.repository;
 
 
 
-import com.Techeer.Team_C.domain.User;
+import com.Techeer.Team_C.User.domain.User;
 
 import java.util.List;
 import java.util.Optional;
